requirements.txt
streamlit>=1.30.0
trafilatura>=1.12.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
google-ads>=24.1.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0
pytest>=7.0.0
